window error

+++ test/tbb/test_concurrent_queue.cpp
@@ -316,18 +316,18 @@ void TestMoveQueue(){
     allocator_type::set_limits(300);
     queue_type q1, q2;
     move_support_tests::Foo obj;
-    int n1(15), n2(7);
+    size_t n1(15), n2(7);
 
     allocator_type::init_counters();
     for(int i =0; i < n1; i++)
       q1.push(obj);
-    int q1_items_constructed = allocator_type::items_constructed;
-    int q1_items_allocated =  allocator_type::items_allocated;
+    size_t q1_items_constructed = allocator_type::items_constructed;
+    size_t q1_items_allocated =  allocator_type::items_allocated;
 
     allocator_type::init_counters();
     for(int i =0; i < n2; i++)
       q2.push(obj);
-    int q2_items_allocated =  allocator_type::items_allocated;
+    size_t q2_items_allocated =  allocator_type::items_allocated;
 
     allocator_type::init_counters();
     q1 = q2;
